add vaarg test for long long type

+++ lbdex/input/ch9_3_template.cpp
@@ -26,7 +26,7 @@ T sum(T amount, ...)
 
 int test_template()
 {
-  int a = sum<int>(6, 0, 1, 2, 3, 4, 5);
+  int a = (int)(sum<int>(6, 0, 1, 2, 3, 4, 5));
 	
   return a;
 }
